Define snapshot strategy for Node
