+++ Tests/AppTests/Util.swift
@@ -36,13 +36,6 @@ func fixturesDirectory(path: String = #file) -> URL {
 }
 
 
-extension XCTestCase {
-    var isRunningInCI: Bool {
-        ProcessInfo.processInfo.environment.keys.contains("GITHUB_WORKFLOW")
-    }
-}
-
-
 // MARK: - Package db helpers
 
 
@@ -94,80 +87,4 @@ func makeBody(_ string: String) -> ByteBuffer {
     var buffer: ByteBuffer = ByteBuffer.init(.init())
     buffer.writeString(string)
     return buffer
-}
-
-
-// MARK: - Useful extensions
-
-
-extension URL: ExpressibleByStringLiteral {
-    public init(stringLiteral value: String) {
-        precondition(!value.isEmpty, "cannot convert empty string to URL")
-        self = URL(string: value)!
-    }
-}
-
-
-extension String {
-    var url: URL { URL(string: self)! }
-}
-
-
-extension Result {
-    var isSuccess: Bool {
-        if case .success = self { return true }
-        return false
-    }
-
-    var isFailure: Bool {
-        if case .failure = self { return true }
-        return false
-    }
-}
-
-
-extension Array where Element: FluentKit.Model {
-    public func save(on database: Database) -> EventLoopFuture<Void> {
-        map {
-            $0.save(on: database)
-        }.flatten(on: database.eventLoop)
-    }
-}
-
-
-extension String {
-    var gh: String { "https://github.com/foo/\(self)" }
-}
-
-
-extension Array where Element == String {
-    var gh: Self { map(\.gh) }
-}
-
-
-// MARK: - custom asserts
-
-
-extension XCTestCase {
-    func assertEquals<Root, Value: Equatable>(_ keyPath: KeyPath<Root, Value>,
-                                              _ value1: Root,
-                                              _ value2: Root,
-                                              file: StaticString = #file,
-                                              line: UInt = #line) {
-        XCTAssertEqual(value1[keyPath: keyPath],
-                       value2[keyPath: keyPath],
-                       "\(value1[keyPath: keyPath]) not equal to \(value2[keyPath: keyPath])",
-                       file: file, line: line)
-    }
-
-    func assertEquals<Root, Value: Equatable>(_ values: [Root],
-                                              _ keyPath: KeyPath<Root, Value>,
-                                              _ expectations: [Value],
-                                              file: StaticString = #file,
-                                              line: UInt = #line) {
-        XCTAssertEqual(values.map { $0[keyPath: keyPath] },
-                       expectations,
-                       "\(values.map { $0[keyPath: keyPath] }) not equal to \(expectations)",
-                       file: file, line: line)
-    }
 }
